HW6/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==1.0.2
Flask-Cors==3.0.10
Flask-JSGlue==0.3.1
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
